server/requirements.txt
websockets
pystray
Pillow
pyinstaller
